+++ test/runtests.jl
@@ -11,5 +11,16 @@ include("linter.jl")
 include("rformula.jl")
 include("kb.jl")
 include("output.jl")
-include("kb_native_linters.jl")
 include("version.jl")
+
+# csv and r
+include("data_csv.jl")
+include("kb_native_linters_csv_r.jl")
+
+# arrow and r
+include("data_arrow.jl")
+include("kb_native_linters_arrow_r.jl")
+
+# parquet and r
+include("data_parquet.jl")
+include("kb_native_linters_parquet_r.jl")